+++ src/test/kotlin/kondorcet/CondorcetMethodTest.kt
@@ -9,12 +9,12 @@ class CondorcetMethodTest {
     fun testCondorcetWinner() {
         val poll = CondorcetMethod<Char>()
 
-        poll.vote(Ballot.of('A', 'C', 'B'), 23)
-        poll.vote(Ballot.of('B', 'C', 'A'), 19)
-        poll.vote(Ballot.of('C', 'B', 'A'), 16)
-        poll.vote(Ballot.of('C', 'A', 'B'), 2)
+        poll.vote(ballot('A', 'C', 'B'), 23)
+        poll.vote(ballot('B', 'C', 'A'), 19)
+        poll.vote(ballot('C', 'B', 'A'), 16)
+        poll.vote(ballot('C', 'A', 'B'), 2)
 
-        val expected = Ballot.of('C', 'B', 'A')
+        val expected = ballot('C', 'B', 'A')
         val actual = poll.result()
 
         Assert.assertEquals(expected, actual)
@@ -29,16 +29,16 @@ class CondorcetMethodTest {
     fun testParadoxSolving() {
         val poll = CondorcetMethod<Char>()
 
-        poll.vote(Ballot.of('A', 'C', 'B', 'E', 'D'), 5)
-        poll.vote(Ballot.of('A', 'D', 'E', 'C', 'B'), 5)
-        poll.vote(Ballot.of('B', 'E', 'D', 'A', 'C'), 8)
-        poll.vote(Ballot.of('C', 'A', 'B', 'E', 'D'), 3)
-        poll.vote(Ballot.of('C', 'A', 'E', 'B', 'D'), 7)
-        poll.vote(Ballot.of('C', 'B', 'A', 'D', 'E'), 2)
-        poll.vote(Ballot.of('D', 'C', 'E', 'B', 'A'), 7)
-        poll.vote(Ballot.of('E', 'B', 'A', 'D', 'C'), 8)
+        poll.vote(ballot('A', 'C', 'B', 'E', 'D'), 5)
+        poll.vote(ballot('A', 'D', 'E', 'C', 'B'), 5)
+        poll.vote(ballot('B', 'E', 'D', 'A', 'C'), 8)
+        poll.vote(ballot('C', 'A', 'B', 'E', 'D'), 3)
+        poll.vote(ballot('C', 'A', 'E', 'B', 'D'), 7)
+        poll.vote(ballot('C', 'B', 'A', 'D', 'E'), 2)
+        poll.vote(ballot('D', 'C', 'E', 'B', 'A'), 7)
+        poll.vote(ballot('E', 'B', 'A', 'D', 'C'), 8)
 
-        val expected = Ballot(listOf(setOf('E', 'A', 'C', 'B', 'D')))
+        val expected = ballot(setOf('E', 'A', 'C', 'B', 'D'))
         val actual = poll.result()
 
         Assert.assertEquals(expected, actual)
@@ -48,10 +48,10 @@ class CondorcetMethodTest {
     fun testExAequoOnTheMiddle() {
         val poll = SchulzeMethod<Char>()
 
-        poll.vote(Ballot.of('A', 'B', 'C', 'D'))
-        poll.vote(Ballot.of('A', 'C', 'B', 'D'))
+        poll.vote(ballot('A', 'B', 'C', 'D'))
+        poll.vote(ballot('A', 'C', 'B', 'D'))
 
-        val expected = Ballot(listOf(setOf('A'), setOf('B', 'C'), setOf('D')))
+        val expected = ballot(setOf('A'), setOf('B', 'C'), setOf('D'))
         val actual = poll.result()
 
         Assert.assertEquals(expected, actual)